+++ tests/test_builtin_recipes.py
@@ -1,5 +1,7 @@
 #!/usr/bin/env python
+import pytest
 from pathlib import Path
+from typing import List, Dict
 
 import alkymi.recipes
 from alkymi import AlkymiConfig
@@ -137,3 +139,70 @@ def test_builtin_kwargs(tmpdir):
     # Changing the file contents back should fix things
     file_a.write_text(file_a.name)
     assert args.recipe.status() == Status.Ok
+
+
+def test_zip_results():
+    """
+    Test that the zip_results() built-in recipe generator works as expected (like zip(), but for recipes)
+    """
+    AlkymiConfig.get().cache = False
+
+    # Test zipping of lists
+    @alkymi.recipe()
+    def numbers() -> List[int]:
+        return [0, 1, 2, 3]
+
+    @alkymi.recipe()
+    def strings() -> List[str]:
+        return ["0", "1", "2", "3"]
+
+    @alkymi.recipe()
+    def spelled() -> List[str]:
+        return ["zero", "one", "two", "three"]
+
+    zipped = alkymi.recipes.zip_results((numbers, strings, spelled))
+    results = zipped.brew()
+    assert isinstance(results, list)
+    assert all(isinstance(result, tuple) for result in results)
+    assert results[0] == (0, "0", "zero")
+    assert results[1] == (1, "1", "one")
+    assert results[2] == (2, "2", "two")
+    assert results[3] == (3, "3", "three")
+
+    @alkymi.recipe()
+    def few_items() -> List[int]:
+        return [42, 42]
+
+    zipped_should_fail = alkymi.recipes.zip_results((numbers, few_items))
+    with pytest.raises(ValueError):
+        zipped_should_fail.brew()
+
+    @alkymi.recipe()
+    def not_iterable() -> int:
+        return 42
+
+    zipped_should_fail_2 = alkymi.recipes.zip_results((numbers, not_iterable))
+    with pytest.raises(ValueError):
+        zipped_should_fail_2.brew()
+
+    # Now try zipping dictionaries
+    @alkymi.recipe()
+    def numbers_dict() -> Dict[str, int]:
+        return dict(zero=0, one=1, two=2, three=3)
+
+    @alkymi.recipe()
+    def strings_dict() -> Dict[str, str]:
+        return dict(zero="0", one="1", two="2", three="3")
+
+    zipped_dict = alkymi.recipes.zip_results((numbers_dict, strings_dict))
+    results_dict = zipped_dict.brew()
+    assert isinstance(results_dict, dict)
+    assert all(isinstance(result, tuple) for result in results)
+    assert results_dict["zero"] == (0, "0")
+    assert results_dict["one"] == (1, "1")
+    assert results_dict["two"] == (2, "2")
+    assert results_dict["three"] == (3, "3")
+
+    # Test that zipping different types fails
+    with pytest.raises(ValueError):
+        alkymi.recipes.zip_results((numbers, numbers_dict)).brew()